Move CodeGenerator to be accessible by other validation tests
Fixes #2342
